pyproject.toml
[tool.poetry]
name = "taskmanager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.1"
PyJWT = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
